+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [13.5.1](https://kaos.sh/ek/13.5.1)
 
+- `[mathutil]` Added method `FromPerc`
 - `[mathutil]` Code refactoring
 
 ### [13.5.0](https://kaos.sh/ek/13.5.0)